server/pyproject.toml
[tool.poetry]
name = "aces-vision-docker"
version = "0.0.1"
description = "Docker API for aces-vision."
authors = ["Kubo Shizuma"]
repository = "https://github.com/aces-inc/aces-vision-docker"

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
acesvision = { git = "ssh://git@github.com/aces-inc/aces-vision.git", tag = "v2.9.1" }
aws_requests_auth = "*"
aws-xray-sdk = "*"
boto3 = "*"
dataclasses-json = "^0.5.7"
filterpy = "*"
Flask = "*"
gunicorn = "*"
matplotlib = "*"
motmetrics = {git = "https://github.com/aces-inc/py-motmetrics.git", tag = "v1.2.1aces"}
numpy = "*"
opencv-contrib-python-headless = "*"
requests = "*"
requests-futures = "*"
requests_testadapter ="*"
pandas = "*"
Pillow = "*"
python-magic = "*"
PyYAML = "*"
sentry-sdk = "*"
scipy = "*"
scikit-image = "*"
scikit-learn = "*"
tensorboard = "*"
tqdm = "*"
vidgear = "*"
pydantic = "^1.10.2"
PyJWT = "*"
xmltodict = "^0.12.0"

[tool.poetry.dev-dependencies]
autopep8 = "*"
flake8 = "*"
flake8-commas = "*"
flake8-isort = "*"
flake8-quotes = "*"
jupyter = "*"
jupyterlab = "*"
pep8-naming = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
